Start refactoring GameObject system

+++ include/render/scale.hpp
@@ -1,16 +1,16 @@
 #pragma once
 
-#include <glm/vec3.hpp>
+#include <glm/glm.hpp>
 
 namespace render
 {
 class Scale
 {
   public:
-    Scale(float scale) : x{scale}, y{scale}, z{scale} {}
+    Scale(double scale) : x{scale}, y{scale}, z{scale} {}
     Scale(glm::vec3 scale) : x{scale.x}, y{scale.y}, z{scale.z} {}
-    const float x;
-    const float y;
-    const float z;
+    const double x;
+    const double y;
+    const double z;
 };
 } // namespace render
